fix(test): update snapshots
