+++ crates/pack-tests/tests/snapshot/basic/target_node/output/main.js
@@ -1,4 +1,4 @@
 var R=require("./_turbopack__runtime.js")("main.js")
-R.c("_project__basic_target_node_input_ae63d396.js")
+R.c("_root-of-the-server___1e3d7703.js")
 R.m("[project]/basic/target_node/input/index.ts [server] (ecmascript)")
 module.exports=R.m("[project]/basic/target_node/input/index.ts [server] (ecmascript)").exports
